Conflicts:
	phoenix-knowledge-api/src/main/java/com/ginkgocap/ywxt/knowledge/service/KnowledgeHomeService.java
	phoenix-knowledge-provider/src/main/java/com/ginkgocap/ywxt/knowledge/mapper/KnowledgeCategoryValueMapper.java
	phoenix-knowledge-provider/src/main/java/com/ginkgocap/ywxt/knowledge/mapper/KnowledgeCategoryValueMapper.xml
	phoenix-knowledge-provider/src/main/java/com/ginkgocap/ywxt/knowledge/service/impl/KnowledgeHomeServiceImpl.java

+++ phoenix-knowledge-provider/src/main/java/com/ginkgocap/ywxt/knowledge/service/impl/KnowledgeDraftServiceImpl.java
@@ -681,6 +681,9 @@ public class KnowledgeDraftServiceImpl implements KnowledgeDraftService {
 			Long userId) {
 
 		Knowledge k = getDraftByMainIdAndUser(knowledgeMainId, type, userId);
+		if(k==null){
+			return 0;
+		}
 		knowledgeNewsDAO.deleteKnowledgeById(k.getId(), type);
 
 		if (userId == null) {
